Use Singleton mixin on middleware chains
Part of the design of Peribot (at least for now) is for each middleware
chain to exist as a single instance. In part, this is because mixing in
Concurrent::Async creates a background thread for each chain to process
messages on (at least according to the concurrent-ruby documentation),
and I want to ensure that resources are being used effectively.

Previously, I did this by including module functions on Peribot's main
module in order to create or access an existing instance of each chain.
However, I don't believe that this is a particularly clean solution as
it unnecessarily clutters up the Peribot module. I think that the
Singleton module provides a more maintainable solution overall.

This is a somewhat large change in the API for middleware chains, so I
have bumped to a new minor version.