Don't run RuboCop during CI builds
Rubinius was segfaulting on Travis while running RuboCop, which caused
the entire build to fail due to RuboCop's fail_on_error option being
enabled. I still wish for this to be enabled locally as I feel that it
is helpful for myself in terms of enforcing proper style, but I don't
want it to break the entire build. So, this change sets up Travis to
only run RSpec. This should make builds a tiny bit faster as well.