Implement Peribot.when_configured
Create the Peribot.when_configured method, which executes a given block
after Peribot has been configured, whether that is now (if configuration
is done) or later (when Peribot.configure gets called).

I updated the spec to reset the @conf_promise that's created when
Peribot.configure gets called. This is horrifying for a number of
reasons, and I'm going to need a better solution in order to improve
this and other tests.