Update spec for Peribot::Middleware::Chain
As part of this refactoring effort, the usage of
Peribot::Middleware::Chain will be changing. Instead of extending this
class in order to create chain classes (with task subclasses created
through complex metaprogramming), chains will be created simply by
instantiating the class and providing an optional end action through a
block.