Don't log Peribot::Middleware::Stop
Ensure that Peribot::Middleware::Stop is not logged when middleware
tasks raise it. This error is intended to simply stop message processing
rather than actually raise an alarm and log a message for debugging
purposes.

Raising an exception really doesn't seem like the "proper" way of doing
this, but it does work well considering how promises are set up. I will
implement a method in the Task class that can be used to stop the
pipeline while hiding this implementation detail.