Create configuration module for bot instances
Implement a module that can be mixed in to provide configuration
facilities for Peribot instances. This will be part of the Peribot::Bot
class.

FUNCTIONALITY IS CURRENTLY DUPLICATED BETWEEN THE MAIN MODULE AND THIS
ONE. IT MUST BE REMOVED FROM THE MAIN MODULE ONCE THE NEW BOT CLASS IS
FULLY IMPLEMENTED.