Remove dependency on GroupMe
Although the overall intention of Peribot is to be used with GroupMe, I
expect that I will include the GroupMe-specific functionality in a
separate gem. Thus, it does not need to be in the Gemfile, which again
cleans up dependencies a bit. Now, the only dependency is
concurrent-ruby, which itself includes no external gem dependencies (as
a design goal).