General cleanup and documentation improvements
Several improvements have been made, particularly to documentation.
Methods have better documentation of return values, and private methods
are explicitly called out more within documentation. The old services
module has been removed, as it was not being used at all and was left in
accidentally. A require statement for Singleton in Middleware::Chain was
removed, as this was left in accidentally as well.

[ci skip]