Bump version to 0.5.1
Release the command matching fix and duplicate handler registration
improvement as 0.5.1 so that it can be used in Peribot instances under
development. Improvements for how listen handlers work when registered
multiple times are in progress and these will likely be released as part
of version 0.6.0 when ready as they will probably involve breaking API
changes.