Ensure that commands in regexes are escaped
Escape commands registered by users before interpolating them into
regexes. Otherwise, commands containing special characters (such as .)
might not properly be recognized and/or ignored in all cases.

In this particular case, testing on a real instance with a command
containing a period showed issues. The period was interpreted as it
normally would be within a regex (as any character), leading to
unexpected results when another character (such as space) was inserted
in its place.