Ensure that listen handlers are only called once
Make sure that no matter how many times a particular handler is matched
by associated regexes, it is only called once. The handler will still
only receive one match data object, which is determined by the first
registered regex that matches for the handler. In other words, priority
is essentially defined by the order in which listen handlers are defined
within a service class.

This is an attempt to enforce a convention that was not previously clear
in Peribot: Handler methods are only called once for a particular
message.