Ensure that command handlers are only called once
Create an additional test to ensure that command handlers are only
called once. As of now, this requires no code changes due to the way
that command handlers are registered. However, this is here to ensure
that future changes do not break the expected behavior (at least unless
the specifications and documentation are changed as well).