Bump version to 0.5.3
This is to account for the updated handler functionality. This is not a
particularly major update, as the overall function signature for listen
handlers is the same. The change simply enforces an invariant that was
intended to exist from the beginning. Thus, I am not creating a new
minor version.

[ci skip]