Clarify when Concurrent::IVars are returned
While processor chains and services run asynchronously, the return
values of #accept for each one provide something that can be waited on
in order to make things behave more synchronously. This is particularly
important for testing.

In this commit, update the tests to call wait on IVars instead of
value. Retrieving the value isn't the goal during testing - it's simply
waiting for processing to finish. This is more-or-less true for other
things that use Peribot components. A service can be waited on, but the
value isn't necessarily guaranteed to be stable.