Add cache functionality to Peribot::Bot
Create a simple cache for Peribot::Bot instances that can be used to
temporarily save values that will not be persisted across runs. This may
be useful for caching data that can change - the cache can be flushed by
restarting the instance.

Future work might involve bringing the store functionality a bit more in
line with this - simply providing access to the store map via
attr_reader.