Ensure processor chains are initialized with bot
Ensure that Peribot::Bot initializes any processor chains it creates
with an instance of Peribot::Bot, as that is the whole point of having
the bot parameter in the constructor in the first place. If the chain is
initialized with something other than this, it may break the assumption
that anything in Peribot::Bot can be used within processor chains.