Improve Peribot::Processor (and tests)
Some minor improvements to Peribot::Processor. In particular, tests are
much better now. The bot is now available through a private accessor
method.

Honestly, I'm not sure that this is really worth a full class. I will
probably just turn this into a module that includes the stop_processing
method. Otherwise, it's not hard to implement a class that takes a
single argument in the constructor and implements #process with one
argument. Preferring duck typing to inheritance, basically.