Minor updates for processor chain and tests
Add some clarifying informational comments to the processor chain to
describe certain decisions that have been made. A minor change was made
to not call #execute on the promise chain (as it is not necessary), and
a minor change was made to sleep for a brief time in the end action (to
ensure that we wait on it properly).