Ensure we never 'break' from promise rescue block
Using 'break' to return early from the rescue block in a promise leads
to an infinite hang on JRuby. I've known this for a while, but had not
mentioned this anywhere in the code or verified through tests that it
was done correctly. Now, there is a test to ensure that we don't break
from whatever is calling the rescue block (which probably had something
to do with the error). This should help enforce proper usage going
forward.