Minor service cleanup and accessor use
Use accessors rather than instance variables in services, per changes
made in other parts of peribot. Additionally, do not explicitly call
execute on the promise created for the message, as it is not necessary
when the promise begins in a fulfilled state (see the processor chain
code for more info).