Allow autoconfig from environment variables
If configuration and store directories are not given explicitly when
creating a bot instance, attempt to pull this automatically from the
PERIBOT_CONFIG_DIR and PERIBOT_STORE_DIR environment variables. This
simplifies explicit configuration requirements for instance maintainers
who wish to use a well-known best practice when managing their
instances.