Create beta version 0.7.0-pre
At this point, I am planning some breaking changes to Peribot. In
particular, the methods for using stores and caches are going to change.
I am also considering the possibility of single-file (rather than
directory-based) configuration and storage.

In order to not disturb current instances, I am performing this work on
a separate branch.

[ci skip]