Allow processors to return nil to stop processing
Rather than forcing processors to raise the
Peribot::ProcessorChain::Stop exception directly in order to stop
further message processing, allow them to return nil. The exception will
then be raised on their behalf.

Effectively, this makes the usage of a special exception to stop the
processor chain an implementation detail that is irrelevant to outside
processors. Considering that it's kind of a funky (even if reasonable
given the circumstances) design decision, this is probably a good thing.