Bump Rubocop to 0.46
Bump Rubocop to the latest version for updates and additional checks.
Note that I've disabled a check for the block hash builder that is used
for configuration. Realistically I should just remove this because it
was kind of a dumb idea that breaks down when any significant
configuration is required.