Update Rubocop to 0.47
Update Rubocop to the latest version, which includes a new block length
metric. I had noticed Syntastic (in Vim) putting out tons of block
length complaints in my tests and got the hint that the bundled version
might have been out of date.

The block length metric is now ignored for tests though. Maybe this can
be changed in the future, maybe not.