Bump versions for Travis CI (and remove rbx)
Bump all of the Ruby versions for Travis CI to the latest available.
Additionally, remove Rubinius from the test list. I allowed failures for
it once they switched to their (stupid) new non-semver versioning
scheme, and now it seems that rvm just doesn't want to install it at
all. So, no point in wasting my time with it I guess.