Use a ProcessorChain for the top-level pipeline
Rather than implement the top-level processing pipeline using recursive
calls to the #accept method, implement it using a ProcessorChain. This
allows for a clearer implementation that leverages one of the major
abstractions introduced by Peribot 0.9.0.

There is a slight change in the behavior of the #accept method as a
result of this change, as the set of processors will be computed at the
beginning of #accept rather than at the beginning of each stage. I
consider the case of registering processors after one has started to
process messages to be unsupported by Peribot, though I'm admittedly not
sure that this is clearly documented at this time.