Document thread-safety and use info for Bot#use
Improve the documentation of Peribot::Bot#use by explaining that it is
designed to be used before all message processing begins, and that it
may be thread-unsafe and that the behavior of pipelines while
registration is being actively performed is not defined. This has always
been the case in Peribot, but this expectation was not explicitly
documented before (or if it was, I likely deleted it during the 0.9.0
changes).