Remove deprecated methods from Bot
With the introduction of registries for all message processing stages,
the register and services methods of Bot became deprecated. They are now
removed in favor of service.register and service.list, which is more
consistent with other processing stages and removes the "special" status
that services held in pre-0.9 Peribot.