Add "filter" as the first processing stage
At the moment, Peribot's preprocessor stage is effectively used to serve
two distinct purposes:

  * Filtering messages that should not be processed
  * Processing (possibly with side effects) certain messages, then
    passing them through

For example, I have a personal Peribot instance that uses a preprocessor
to send a "welcome" message the very first time Peribot is added to a
group. However, Peribot::GroupMe in bot sending mode uses a preprocessor
to filter out messages not destined for the bot. If someone were to
implement a similar feature, use bot sending mode with Peribot::GroupMe,
and register the two components in that order, it is possible that a
welcome message would be sent even when it should have been filtered.

The idea of the "filter" and "preprocessor" stages is that they each
handle one of the two distinct cases described above. A filter should
ONLY pass or drop a message without side effects based on the state and
configuration of the bot, while a preprocessor could be more flexible
and execute side effects resulting from a message.

Filtering is implemented as a ProcessorChain, just like preprocessing
and postprocessing. Notably, the refactoring introduced for 0.9 made
this change extremely simple (in fact, without tests the change would
have only been one line).