Add initial changelog
Personally, I've come to hate it when projects don't include a
changelog. It makes it way harder for me to know whether it's safe to
upgrade my dependencies.

I've decided it's long past time to practice what I preach, thus I'm
going to keep a changelog for Peribot as well.