Refactor Service to stop using so many promises
This is a cleanup that I've been wanting to do for a long time, since
the promise-based logic was getting annoying to deal with. I was
initially hesitant about breaking backwards-compatibility by not using
Concurrent::Promise, but then realized I could just use it at the top
level instead of everywhere in the service processing chain. Make sense?

One potentially controversial move is to prefix all of the private
method names with underscores. The goal is for them to never conflict
with names in subclasses, so it's unlikely that I'd take this away even
though it looks ugly.