Fix "should not have output" service tests
While using Mutant to help validate the recent Peribot::Service
refactoring, I found that one of my unit tests wasn't killing a mutant.
I had forgotten that because everything executes in the context of a
Promise, an exception in the sender would *not* actually throw an
exception, but instead reject the promise!

This is solved with explicit checks on the "fulfilled" states of these
promises. I've verified that this form actually does fail when it needs
to.